Buisness rule cant delete flight id 1and2

+++ src/main/java/com/qa/buisiness/service/FlightServiceImpl.java
@@ -55,7 +55,12 @@ public class FlightServiceImpl implements FlightService {
 	}
 
 	public String deleteFlight(Long flightId) {
-		return repo.deleteFlight(flightId);
+
+		if (flightId <= 2) {
+			return "{\"message\": \"Cannot Delete Flight\"}";
+		} else {
+			return repo.deleteFlight(flightId);
+		}
 	}
 
 	// flight destinations
